[TID] Replace circular area estimate with summed triangles.
Instead of assuming the cell is roughly spherical we now sum the area of
the triangles formed by each piecewise pair of boundary elements and the
cell centre.

Also add a mechanism to move the boundaries of the simulation over time
to observe the cell under compression.  Behaves very nicely, much better
than the old model.